Add condition operators for working with regexes.
This introduces two new operators for matching or not matching against a
user-provided regex. The regex is checked for validity and stored as a
serialized pattern in the condition, and then used to match the given
string during the condition matching process.

`BaseConditionStringChoice` now includes these by default. It also adds
the "Does Not Contain" operator, which was missing before.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8346/