Add a template for rendering legacy and Django fieldsets.
This is a combination of two backports, one for the new fieldsets
template, and one for adding support for Django fieldsets.

The new fieldsets template is available in
`siteconfig/settings_fieldsets.html`, unlike the new template which is in
`forms/admin/form_fieldsets.html`. The legacy location is used to
prevent breakages with users of Djblets, since it would require adding
the `djblets.forms` to `INSTALLED_APPS`.

The old template that this fieldset support came from only supported
our old custom version of fieldsets, rather than Django's version.
In Djblets 0.10, we added support for both. This change is backporting
that to the new fieldsets template.

Testing Done:
Tested with old-style and new-style fieldsets.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8376/