Add utility functions for building e-mail addresses.
This adds a couple of small utility functions that take a user (or a
user's information) and constructs an e-mail address intended for use in
a To/CC/BCC field. These are ports from the existing e-mail code that
lives in Review Board, and will be used and improved upon in some
upcoming e-mail work in Djblets.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8398/