Revert the change to combine from_email and extra_headers['From'].
In an earlier change (2c4310d), I removed the separation of the
from_email attribute and custom 'From' header, as it didn't appear
necessary for the e-mail behavior we had. It turns out, however, that
this change was important after-all. Deep in the bowels of Django's
e-mail code, these two values are treated differently. This part of the
change has been reverted, with a new comment added explaining the
difference.

Testing Done:
Updated unit tests for the new expected values. They pass.

Reviewed at https://reviews.reviewboard.org/r/8413/