Disable smart e-mail spoofing in the test runner.
The new smart e-mail spoofing is great in production, but bad in
testing. By default, we want to turn this off for unit tests, forcing
them to opt-in (or override this test runner behavior) to keep tests
predictable and fast.

This change was already reviewed on the Review Board side. It's simply
and trivially being ported over to Djblets for 0.10.