Consume a SpyAgency for DmarcDnsTestsMixin instead of inheriting.
The new DmarcDnsTestsMixin subclassed SpyAgency, which ended up causing
a chain reaction of problems from subclasses that needed to define their
own SpyAgency. Untangling that was a hairy mess, all for one little spy.

This is a trivial change that simply makes use of an internal SpyAgency
within the mixin instead of subclassing one.