Add the URLAvatarService
The `URLAvatarService` is a new avatar service for setting the avatar of
a user to an absolute URL. This service is a hidden service, which is a
new type of service that is not exposed to users through the
configuration form. (Indeed, attempting to fool the form and submitting
the data will result in an error as if it does not exist.) Instead, it
is to be used to provide avatars for users created by extensions, etc.

Testing Done:
- Ran unit tests.
- Tested with Review Board.
- Saw that the service did not show up in the avatar services form.

Reviewed at https://reviews.reviewboard.org/r/8481/