Fix packaging on Python 2.6.
The new and improved setup.py support fails on Python 2.6, due to the
inability to directly call `python -m pip`. Instead, we have to specify
the specific module by calling `python -m pip.__main__`. This is not
required for Python 2.7, but works fine there.

This change will not be merged to release-0.10.x.

Testing Done:
Built the package successfully on both Python 2.6 and 2.7.

Reviewed at https://reviews.reviewboard.org/r/8503/