Send an X-CSRFToken header when deleting integrations
Previously, integrations could not be deleted. The view required a CSRF
token to be submitted but we were not sending one. Now we check for a
CSRF token field in the form before executing our `HTTP DELETE` request
and, if present, we add the `X-CSRFToken` header to the request. This
allows integrations to be deleted.

Testing Done:
Able to delete an integration with this patch applied.

Reviewed at https://reviews.reviewboard.org/r/8490/