Install the Djblets-specified npm dependencies for extension packaging.
This updates the new npm installation code using during extension
packaging to instead make use of the npm dependency lists used for
Djblets itself. This ensures compatibility/consistency between Djblets
media and extension media.

The dependencies are now specified in a package.json file, which is
stored in the build/ directory. This allows for a quicker
installation/checking of npm dependencies. The resulting node_modules/
also goes in build/, in order to keep the root of the extension tree
clean.

This will allow for better maintenance of extensions and dependencies
going forward, and allow us to more easily provide the same types of
less/babel plugins we depend on ourselves with minimal maintenance.

Testing Done:
Tested with an extension containing `.less` files, `.js`, and `.es6.js`.
Built the package, saw it install the modules and then compile the files.
Checked the build output and saw all the correct file contents.

Reviewed at https://reviews.reviewboard.org/r/8507/