Check for node and nodejs when running setup.py develop.
If node isn't installed, `setup.py develop` will fail part-way through
the installation process. However, if it's installed but as `nodejs`
(such as on Ubuntu), it will all work correctly until you go to run
Review Board, at which point it will fail confusingly when it fails to
find `node` in the path.

We now check for both `node` and `nodejs`, and provide appropriate error
messages depending on what we find or don't find.

Testing Done:
Tested with `node` in the path and without.

Tested without `node` in the path but with and without `nodejs` in the
path.

Saw the expected behavior in all cases.

Reviewed at https://reviews.reviewboard.org/r/8511/