Fix a regression in the test suite from the extension doc changes.
The extension documentation changes removed an old internal method that
was used for a property, in favor of using the property decorators
instead. One of our unit tests was accessing the internal method instead
of the property, though, which caused tests to break. This changes the
test to use the property instead.