Add ETag data generation to the avatar service API
Avatar services can now generate ETag (Entity Tag) data to be used in
generating caching information for the avatars. This will allow pages
that use the avatar to be cached and only retrieved if the avatar
changes.

Testing Done:
Tested this with Review Board in /r/8567. Saw that it worked correctly.

Reviewed at https://reviews.reviewboard.org/r/8566/