Require django-pipeline 1.6.9.
The latest version of django-pipline is 1.6.10, but this version has a
bug that causes every single static media file to be rebuilt on every
page load. 1.6.11 should contain the fix, but for now we want to avoid
this version.

This is a trivial change that hard-codes 1.6.9 as the required version
(updating from a minimum version of 1.6.8) and adds a note as to why
we've pinned this to 1.6.9.