Add anchors, intersphinx mappings, and error fixes for Djblets docs.
This introduces a couple of anchors needed to easily reference the
Djblets guides and code reference pages. These are used by Review Board
today, but these corresponding changes never made it into Djblets.

There's also an intersphinx mapping for setuptools that's used to link
to code and documentation.

And finally, some fixes for errors reported when building the docs. We
had some docstrings that were not ReST-friendly, due to indentation or
bad syntax for references. This fixes up those docs to be error-free.

Testing Done:
Built the docs successfully. No errors reported.

Verified that the Review Board docs could link to the correct pages in
the Djblets docs.

Reviewed at https://reviews.reviewboard.org/r/8668/