Add utility functions for clearing template-related caches.
This adds some useful functions for clearing the template caches in the
loaders and the template tag library caches. This supports the various
caching methods used in Django 1.6 and higher.

We'll be able to use this in the extension manager to safely handle
cache clearing for all supported versions of Django.

Unit tests were added that ensure this functionality works for all
versions.

Testing Done:
Unit tests pass for Django 1.6, 1.7, 1.8, 1.9, and 1.10.

Reviewed at https://reviews.reviewboard.org/r/8703/