Ensure form registry is populated when adding or removing forms to pages
The `ConfigPageRegistry` was not populating the `ConfigPageFormRegistry`
when calling `add_form_to_page` or `remove_form_from_page`, which would
lead to issues when removing pages in tbe case that the method
was called before the registry was populated. We now specifically
populate the form registry to avoid this.

Testing Done:
Ran unit tests. The new unit tests fail without this fix applied.

Reviewed at https://reviews.reviewboard.org/r/8725/