Instantiate avatar services form when there is no default service
Previously, when the avatar service form was instantiated, it assumed
that there was a default avatar service. If there was not, the form
would fail in instantiation. This patch addresses this.

Testing Done:
Ran unit tests. The provided test fails without the fix applied.

Reviewed at https://reviews.reviewboard.org/r/8728/