Fix looking up static media needed for Condition value fields.
`ConditionsWidget.media` attempted to look up media files needed for all
widgets that comprised the input for the conditions. This didn't
actually work in practice, and failed silently on template load. There
were a couple of problems, including the assumption that the default
value field for a choice was not dynamic, and that all value fields had
form widgets.

This fixes up that code, allowing media to be properly loaded for all
fields in the widget.

Testing Done:
Tested along with some other changes that made use of widget media.
Verified that this was no longer crashing and that the media was able
to be returned for the page.

Reviewed at https://reviews.reviewboard.org/r/8788/