Fix the integration and extension pages on default Django setups.
The integration and extension pages weren't styled in a way that worked
in the default Django administration UI. There are CSS classes we should
have been working with that would handle some of the important styling
for us, but we were ignoring those and doing the work ourselves, which
only ended up looking correct in Review Board.

The integration listings template in particular was very wrong,
inheriting from the admin change page template, which brought with it
some unwanted static media, elements, and class names.

The extension management scripts were also assuming the presence of a
SITE_ROOT variable, which is defined in Review Board. It also assumed
the API would be in a certain location. Now the API URL is looked up in
the template and passed to the JavaScript, which handles it through
normal Backbone means.

This ends up fixing these pages on Review Board and on default Django
installs, and will help with future administration UI styling work.

Testing Done:
Tested the extension and integration pages on Review Board and on a
plain Django administration UI.

Reviewed at https://reviews.reviewboard.org/r/8791/