Correctly compute uploaded avatar hash
The file upload avatar service was incorrectly computing the file hash
for use in ETag computations, but only if the `file_hash` was not
already cached in the user's settings. This issue has been fixed.

Testing Done:
Uploaded an avatar and deleted the `file_cache` key. It caused an
`AttributeError`. After applying this patch, the error would not
reproduce.

Reviewed at https://reviews.reviewboard.org/r/8840/