Ensure avatar service configuration forms are valid
It was previously possible to save the `AvatarSettingsForm` with invalid
form data for a avatar service configuration form. This was because we
only cleaned the form, but we did not do anything with the result. We
now ensure that the avatar service's configuration is valid, otherwise
we will make the `AvatarSettingsForm` invalid.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8859/